Changed pagination helper function inside of the paginatedResponse changed the constructor to get a Panator object instead of the array
